feat: triangle backtest service + fix Hibernate BigDecimal on trade entities
- Add TriangleBacktestService: detects ASCENDING/DESCENDING/SYMMETRIC triangles
  from 5 pivots on 1h bars, C1 (trendline touch + reversal) and C2 (break+retest)
  entries, EDT detection via MACD divergence, ATR-based quality filter
- Add runRelianceTriangleBacktest test: 17 setups, 76.5% WR, C1=91.7% WR
- Fix TradeSignal.dailyRsi and stochRsiK: Double → BigDecimal to satisfy
  Hibernate 6 constraint that scale has no meaning for SQL floating point types

+++ src/main/java/com/dtech/kitecon/trade/entity/TradeSignal.java
@@ -69,10 +69,10 @@ public class TradeSignal {
     private Boolean rsiDivergence;
 
     @Column(precision = 6, scale = 2)
-    private Double dailyRsi;
+    private BigDecimal dailyRsi;
 
     @Column(precision = 8, scale = 4)
-    private Double stochRsiK;
+    private BigDecimal stochRsiK;
 
     @Column(length = 10)
     private String timeframe;
